implement snapshot function
PCS support clone-to-template operation, which clones existing container
to a container of special type. It, in turn, can be cloned later and the
new container will be new auto-generated mac addresses on network
adapter, new ID, new name, e.t.c.

Image in glance is an array of data and metadata on form of key:value
pairs. So we can't upload directory with container as-is. Let's use tar
to convert directory with several files into one data stream.

Please, note, that you have to add following line to a rootwrap.conf
file:

tar: CommandFilter, /bin/tar, root

It a security flaw, but I'll fix it later.