remove network adapters from VE before saving image
We don't need to store information about network adapters in snapshot.
Moreover disk data is the only thing we need to store in snapshot, but
it's not possible with parallels SDK to create container from existing
disk.

We will add new network adapters each time in spawn method. So
snapshot method is a good place to clean network configuration.