add basic support of VMs
Add support of running VMs as openstack instances. You have to create
image in glance with container format pcs-vm and then launch it,
proper type of virtualization will be automatically selected.

Netwokirng is not supported yet, because vzctl is used to configure
network in container and that place have to be fixed before enabling
networking in VMs.