implement networking with help of prlsdkapi
We can't use vzctl for VMs, so it's needed to manage network in VEs
using prlsdkapi.

It's not possible to get device in host with given name using only
prlsdkapi, so I use 2 hacks:

1. If you specify unexistent virtual network id, device will not be
   connected to any bridges, and VE will be started correctly.
2. Interface can be renamed if it's down.

So all works exactly as before: we create device in VE, connect it to
linux bridge or OVS bridge and then enable DHCP on this device.

Unfortunately, this code still will not work for VMs, because
plug method need existing device, but if we start VM, network device
can't be added to it.