add firewall class for neutron
Here is how neutron with OVS works:
br-int <pcsvoXXX>|--|<pcsvbXXX> pcsbrXXX <veth100.0>|---<eth0 in CT>

First, we have integration bridge br-int and veth pair, one
end of which is inside container, another one (veth100.0) -
in host. We have to connect veth100.0 to bridge br-int.

Also neutron need to setup security group rules between
container and br-int. IptablesFirewallDriver class does it
with help of iptables' match physdev:
PHYSDEV match --physdev-out veth100.0 --physdev-is-bridged
PHYSDEV match --physdev-in veth100.0 --physdev-is-bridged

The problem is that this match works only for devices,
connected to a linux bridge, not OVS. So we connect
veth100.0 to br-int through intermediate linux bridge
pcsbrXXX.

There is OVSHybridIptablesFirewallDriver, but we can't use
it, because it assumes, that name of the network device
is the one, which nova generated - tapXXXX instead of vethXXX.X.

We can use base class IptablesFirewallDriver, but we
need to provide _get_device_name method, which returns
a device name for physdev match. We use external_ids in
openvswitch port to pass device name from pcs nova driver
to this class. So here we just find OVS port and return
external_ids.iface-name from it.