implement setting admin password
First, admin password comes to spawn method, so we need
to set it there. And there is also additional method
set_admin_passwd.

It should work for containers in any case, for VMs - only
if it's linux and has parallels tools installed. But now
it doesn't work for VM, because waiting for system boot is
broken for VMs. We need to fix that problem first.