store images in different formats in pcs-glance-tool
Add ability to store images in different formats (which
qemu-img supports) to pcs-glance-tool.

If we store images in formats, which openstack supports from
scratch, them these image can be converted to volumes without
any effort, while converting image in ploop format to a volume
on different storage requires at least support of ploop in
qemu-img and probably modification of cinder code.

To upload image in a foreign format we simple mount ploop and
convert it to another format with qemu-img convert.