sync VE power state in plug_vifs
Sometimes when we reboot host or restart compute service
VE states becomes not the one which expected by Openstack,
so we need to set VE to the state writter in openstack.
plug_vifs method seems to be good place for it, because
compute_manager calls it for each instance on service start.