template: introduce CachedImage class
Add class for handling cached images. Work for retrieving
images from cache is logically separated from creating VMs,
so move this code to separate class, which is not a child of
PCSTemplate.

Class CachedImage has one method - put_image, which downloads
an image, if it's not in cache and then copies image to the
specified destination.

Currently only cache of tarred and compressed with prlcompress
images is supported, so there is only one subclass of CachedImage -
LZRWCachedImage.

This patch doesn't change behaviour, just sort of refactoring.