handle unexpected power state change correctly
If power state of VE is change from outside the OpenStack,
it handles it with some logic: if VE is unexpectedly running
or suspended - it calls power_off method. If it is unexpectedly
running - it also calls power_off. We don't know, how if this
logic will be changed or not. So let's always set requested
state without making assumptions about current state.